make a clearer distinction between reading and writing paths for spaces
