+++ lib/spaces/spaces/models/paths.rb
@@ -1,0 +1,21 @@
+module Spaces
+  module Paths
+
+    def path_for(identifiable)
+      path.join(identifiable.context_identifier)
+    end
+
+    def path
+      universe.path.join(identifier)
+    end
+
+    def file_names_for(directory, identifier)
+      file_path_for(directory, identifier).glob('**/*').reject(&:directory?)
+    end
+
+    def file_path_for(symbol, context_identifier)
+      path.join("#{context_identifier}", "#{symbol}")
+    end
+
+  end
+end
